requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
twilio
python-dotenv
deta